+++ MarketService/Models/ProductModel.cs
@@ -1,6 +1,5 @@
 using System.ComponentModel.DataAnnotations;
 using Libplanet;
-using MarketService.Response;
 using MarketService.Response.Interface;
 using Microsoft.EntityFrameworkCore;
 
@@ -13,8 +12,8 @@ public class ProductModel : IProductSchema
 
     public Address SellerAgentAddress { get; set; }
     public Address SellerAvatarAddress { get; set; }
-    public decimal Price { get; set; }
-    public decimal Quantity { get; set; }
+    public int Price { get; set; }
+    public int Quantity { get; set; }
     public long RegisteredBlockIndex { get; set; }
     public bool Exist { get; set; }
     public bool Legacy { get; set; }
